naive paralellization

+++ assignment2/Poisson3D/jacobi.h
@@ -6,6 +6,6 @@
 #ifndef _JACOBI_H
 #define _JACOBI_H
 
-int jacobi(double ***u, double ***old_u, double ***f, const int N, const int iter_max, const double threshold);
+double jacobi(double ***u, double ***old_u, double ***f, const int N, const int iter_max, const double threshold);
 
 #endif
